Rename frames.py to tables.py
